📖 DOC: 👌 IMPROVE:  extracted utils func add comments for use case
